[BUGFIX] Usage of deprecated t3lib_div::int_from_ver()
The deprecated t3lib_div:int_from_ver() is still called in
ext_localconf.php for every request to the TYPO3 website.
Besides that, this leads to a fatal error since the method
has been removed in TYPO3 6.1 from GeneralUtility, which has
been the namespace pendant to t3lib_div since TYPO3 6.0.

The solution is to invoke the compatibility class that will
determine whether TYPO3 >= 6.0 is used.

Change-Id: I486e6a03de7a98f38dd13ac0131354e53bb795b0
Fixes: #46313
Reviewed-on: https://review.typo3.org/19531
Tested-by: Ernesto Baschny
Tested-by: Myroslav Holyak
Tested-by: Markus Günther
Tested-by: Tomita Militaru
Tested-by: Daniel Hürtgen
Reviewed-by: Ernesto Baschny
Reviewed-by: Myroslav Holyak
Reviewed-by: Markus Günther
Reviewed-by: Tomita Militaru
Reviewed-by: Daniel Hürtgen
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader